requirements.txt
streamlit
google-generativeai
python-dotenv
langchain #read pdf
pyPDF2
faiss-cpu
langchain_google_genai